requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==2.2.3
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
cryptography==44.0.1
debugpy==1.8.12
decorator==5.2.1
executing==2.2.0
filelock==3.17.0
h11==0.14.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
stack-data==0.6.3
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0